ApiSales. sales_data configured.

+++ database/migrations/2021_04_17_230323_create_sales_table.php
@@ -18,12 +18,12 @@ class CreateSalesTable extends Migration
             $table->integer('product_id');
             $table->integer('store_id');
             $table->integer('qty');
-            $table->integer('price');
+            $table->double('price');
             $table->integer('customer_id')->nullable();
             $table->integer('payment_type')->nullable();
-            $table->integer('paid')->nullable();
-            $table->integer('received')->nullable();
-            $table->integer('change')->nullable();
+            $table->double('paid')->nullable();
+            $table->double('received')->nullable();
+            $table->double('change')->nullable();
             $table->integer('order_id')->nullable();
             $table->integer('user_id')->nullable();
             $table->integer('is_offer')->nullable();
